añadir y eliminar Animales

+++ views/animals/index.php
@@ -1,4 +1,5 @@
 <div>
+<a class="btn btn-success btn-share" href="<?php echo ROOT_PATH; ?>animals/addAnimal">Add Animal</a>
 
 <?php foreach ($viewmodel[0] as $item0) :?><!-- Bucle para recorer los animales y dentro de el los comentarios -->
 	<div class="well">
@@ -6,8 +7,15 @@
 
 			<?php if(isset($_SESSION['is_logged_in'])) : ?>
 			<a class="btn btn-success btn-share" href="<?php echo ROOT_PATH; ?>animals/add/<?php echo $item0['idAnimal']?>">Comment</a>
+			
+			
+			<!-- <a class="btn btn-success btn-share" href="<?php echo ROOT_PATH; ?>animals/acoger/<?php echo $item0['idAnimal']?>">Acoger</a> -->
+			
+			
+			<a class="btn btn-success btn-share" href="<?php echo ROOT_PATH; ?>animals/adoptar/<?php echo $item0['idAnimal']?>">Adoptar</a>
 			<?php endif; ?>
-		<?php foreach($viewmodel[1] as $item1) : ?>
+
+		<?php foreach($viewmodel[1] as $item1) : ?><!-- Bucle de los comentarios -->
 			<?php if($item0['idAnimal'] == $item1['idAnimal']) : ?>
 			<div class="well">
 				<h3><?php echo $item1['titulo']; ?></h3>
@@ -15,11 +23,18 @@
 				<hr />
 				<p><?php echo $item1['texto']; ?></p>
 				<br>
-				<a class="btn btn-danger"href="animals/delete/<?php echo $item1['idComentario']?>">Delete Comment</a>
+				<?php if(isset($_SESSION['is_logged_in'] ) ) : ?>
+					<?php if($item1['idUsuario'] == $_SESSION['user_data']['idUsuario']) : ?>
+						<a class="btn btn-danger"href="animals/delete/<?php echo $item1['idComentario']?>">Delete Comment</a>
+					<?php endif; ?>
+				<?php endif; ?>
 			</div>
 			<?php endif; ?>
 
-		<?php endforeach; ?>
+		<?php endforeach; ?><!-- Fin bucle de los comentarios -->
+		<br>
+		<a class="btn btn-danger"href="animals/deleteAnimal/<?php echo $item0['idAnimal']?>">Delete Animal</a>
+
 	</div>
 <?php endforeach; ?><!-- Fin del bucle de los animales -->
